fix(win32): use wx/wx.h and WIN32_LEAN_AND_MEAN to resolve winsock conflicts

+++ src/mswin/WindowsMidiInterface.cpp
@@ -20,12 +20,11 @@
 // Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 //*****************************************************************************
 
-#include <windows.h>
+#include <wx/wx.h>
 #include <mmsystem.h>
 #include <memory.h>
 
 #include "WindowsMidiInterface.h"
-
 #include "WindowsAudioInterface.h"
 
 extern "C"
@@ -236,7 +235,7 @@ void CALLBACK midiIntTimerHandler(
   JZMidiEvent* pMidiEvent = pState->play_buffer.peek();
   while (pMidiEvent)
   {
-    if (pMidiEvent->ref > pState->play_time)
+    if (pMidiEvent->ref > (DWORD)pState->play_time)
     {
       break;
     }
@@ -621,7 +620,7 @@ void CALLBACK midiMtcTimerHandler(
   JZMidiEvent* pMidiEvent = pState->play_buffer.peek();
   while (pMidiEvent)
   {
-    if (pMidiEvent->ref > pState->play_time)
+    if (pMidiEvent->ref > (DWORD)pState->play_time)
     {
       break;
     }
